remove utils.py
