+++ mqboost/dataset.py
@@ -1,3 +1,4 @@
+from itertools import chain, repeat
 from typing import Callable
 
 import lightgbm as lgb
@@ -11,8 +12,73 @@ from mqboost.base import (
     FittingException,
     ModelName,
     TypeName,
+    ValidationException,
 )
-from mqboost.utils import alpha_validate, prepare_x, prepare_y, to_dataframe
+
+
+def validate_alpha(
+    alphas: list[float] | float,
+) -> list[float]:
+    """Validates the list of alphas ensuring they are in ascending order and contain no duplicates."""
+    if isinstance(alphas, float):
+        alphas = [alphas]
+
+    if not isinstance(alphas, list):
+        raise ValidationException("Alpha must be a list or float")
+
+    if 0.0 in alphas or 1.0 in alphas:
+        raise ValidationException("Alpha cannot be 0 or 1")
+
+    _len_alphas = len(alphas)
+    if _len_alphas == 0:
+        raise ValidationException("Input alpha is not valid")
+
+    if _len_alphas >= 2 and any(
+        alphas[i] > alphas[i + 1] for i in range(_len_alphas - 1)
+    ):
+        raise ValidationException("Alpha is not ascending order")
+
+    if _len_alphas != len(set(alphas)):
+        raise ValidationException("Duplicated alpha exists")
+
+    return alphas
+
+
+def prepare_x(
+    x: pd.DataFrame,
+    alphas: list[float],
+) -> pd.DataFrame:
+    """Prepares and returns a stacked DataFrame of features repeated for each alpha, with an additional column indicating the alpha value.
+    Raises:
+        ValidationException: If the input data contains a column named '_tau'.
+    """
+    if "_tau" in x.columns:
+        raise ValidationException("Column name '_tau' is not allowed.")
+
+    _alpha_repeat_count_list = [list(repeat(alpha, len(x))) for alpha in alphas]
+    _alpha_repeat_list = list(chain.from_iterable(_alpha_repeat_count_list))
+
+    _repeated_x = pd.concat([x] * len(alphas), axis=0).reset_index(drop=True)
+    _repeated_x = _repeated_x.assign(
+        _tau=_alpha_repeat_list,
+    )
+    return _repeated_x
+
+
+def prepare_y(
+    y: pd.Series | npt.NDArray,
+    alphas: list[float],
+) -> npt.NDArray:
+    """Prepares and returns a stacked array of target values repeated for each alpha."""
+    return np.concatenate(list(repeat(y, len(alphas))))
+
+
+def to_dataframe(x: pd.DataFrame | pd.Series | npt.NDArray) -> pd.DataFrame:
+    if isinstance(x, np.ndarray) or isinstance(x, pd.Series):
+        _x = pd.DataFrame(x.copy())
+    else:
+        _x = x.copy()
+    return _x
 
 
 class MQDataset:
@@ -40,7 +106,7 @@ class MQDataset:
         """Initialize the MQDataset."""
         self._model = ModelName[model]
         self.nrow = len(data)
-        self.alphas = alpha_validate(alphas)
+        self.alphas = validate_alpha(alphas)
 
         _funcs = FUNC_TYPE[self._model]
         self.train_dtype = _funcs[TypeName.train_dtype]
@@ -68,7 +134,7 @@ class MQDataset:
     def label_mean(self) -> float:
         """Get the label mean."""
         self.__label_available()
-        return self._label_mean
+        return float(self._label_mean)
 
     @property
     def weight(self) -> npt.NDArray | None:
