+++ tests/test_dataset.py
@@ -9,7 +9,152 @@ from mqboost.base import (
     ModelName,
     ValidationException,
 )
-from mqboost.dataset import MQDataset
+from mqboost.dataset import (
+    MQDataset,
+    prepare_x,
+    prepare_y,
+    to_dataframe,
+    validate_alpha,
+)
+
+
+# Test for to_dataframe
+def test_to_dataframe_with_dataframe():
+    x = pd.DataFrame(
+        {
+            "feature_1": [1, 2],
+            "feature_2": [3, 4],
+        }
+    )
+    pd.testing.assert_frame_equal(x, to_dataframe(x))
+
+
+def test_to_dataframe_with_series():
+    x = pd.Series([1, 2, 3])
+    expected = pd.DataFrame(
+        {
+            0: [1, 2, 3],
+        }
+    )
+    pd.testing.assert_frame_equal(expected, to_dataframe(x))
+
+
+def test_to_dataframe_with_array():
+    x = np.array([[1, 2], [3, 4]])
+    expected = pd.DataFrame(
+        {
+            0: [1, 3],
+            1: [2, 4],
+        }
+    )
+    pd.testing.assert_frame_equal(expected, to_dataframe(x))
+
+
+# Test for prepare_x
+def test_prepare_x_with_dataframe():
+    x = pd.DataFrame(
+        {
+            "feature_1": [1, 2],
+            "feature_2": [3, 4],
+        }
+    )
+    alphas = [0.1, 0.2]
+    result = prepare_x(x, alphas)
+    expected = pd.DataFrame(
+        {
+            "feature_1": [1, 2, 1, 2],
+            "feature_2": [3, 4, 3, 4],
+            "_tau": [0.1, 0.1, 0.2, 0.2],
+        }
+    )
+
+    pd.testing.assert_frame_equal(result, expected)
+
+
+def test_prepare_x_with_series():
+    x = pd.Series([1, 2, 3])
+    alphas = [0.1, 0.2]
+    result = prepare_x(to_dataframe(x), alphas)
+    expected = pd.DataFrame(
+        {
+            0: [1, 2, 3, 1, 2, 3],
+            "_tau": [0.1, 0.1, 0.1, 0.2, 0.2, 0.2],
+        }
+    )
+    pd.testing.assert_frame_equal(result, expected)
+
+
+def test_prepare_x_with_array():
+    x = np.array([[1, 2], [3, 4]])
+    alphas = [0.1, 0.2]
+    result = prepare_x(to_dataframe(x), alphas)
+    expected = pd.DataFrame(
+        {
+            0: [1, 3, 1, 3],
+            1: [2, 4, 2, 4],
+            "_tau": [0.1, 0.1, 0.2, 0.2],
+        }
+    )
+    pd.testing.assert_frame_equal(result, expected)
+
+
+def test_prepare_x_raises_on_invalid_column_name():
+    x = pd.DataFrame({"_tau": [1, 2], "feature_1": [3, 4]})
+    alphas = [0.1, 0.2]
+    with pytest.raises(ValidationException, match="Column name '_tau' is not allowed."):
+        prepare_x(x, alphas)
+
+
+# Test for prepare_y
+def test_prepare_y_with_array():
+    y = np.array([1, 2, 3])
+    alphas = [0.1, 0.2]
+    result = prepare_y(y, alphas)
+    expected = np.array([1, 2, 3, 1, 2, 3])
+    np.testing.assert_array_equal(result, expected)
+
+
+def test_prepare_y_with_series():
+    y = pd.Series([1, 2, 3])
+    alphas = [0.1, 0.2]
+    result = prepare_y(y, alphas)
+    expected = np.array([1, 2, 3, 1, 2, 3])
+    np.testing.assert_array_equal(result, expected)
+
+
+# Test for validate_alpha
+def test_validate_alpha_single_alpha():
+    alphas = 0.3
+    result = validate_alpha(alphas)
+    assert result == [0.3]
+
+
+def test_validate_alpha_multiple_alphas():
+    alphas = [0.1, 0.2, 0.3]
+    result = validate_alpha(alphas)
+    assert result == alphas
+
+
+def test_validate_alpha_raises_on_zero_or_one_alpha():
+    with pytest.raises(ValidationException, match="Alpha cannot be 0 or 1"):
+        validate_alpha([0.0, 0.3])
+    with pytest.raises(ValidationException, match="Alpha cannot be 0 or 1"):
+        validate_alpha([0.3, 1.0])
+
+
+def test_validate_alpha_raises_on_non_ascending_alphas():
+    with pytest.raises(ValidationException, match="Alpha is not ascending order"):
+        validate_alpha([0.3, 0.2, 0.1])
+
+
+def test_validate_alpha_raises_on_duplicate_alphas():
+    with pytest.raises(ValidationException, match="Duplicated alpha exists"):
+        validate_alpha([0.1, 0.2, 0.2])
+
+
+def test_validate_alpha_raises_on_empty_alphas():
+    with pytest.raises(ValidationException, match="Input alpha is not valid"):
+        validate_alpha([])
 
 
 def _concat(df: pd.DataFrame, concat_count: int):
